Add files via upload
Assignment 2 :- I have tried to Register the Olay website,but it is getting Error of Zip/postal code. 
I tried with different postal code but can't able to login. I am considering this as a defect.

Assignment 3:- Makemy Trip 
I have written the code but the Xpath are not getting executed. I am getting Error is no such element found and google it also but can't rectified the error.